requirements.txt
asttokens==2.4.1
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
cssbeautifier==1.15.1
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
djlint==1.34.1
dnspython==2.5.0
EditorConfig==0.12.4
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
Flask==3.0.2
Flask-Limiter==3.7.0
Flask-WTF==1.2.1
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.7
importlib_resources==6.4.0
ipykernel==6.29.3
ipython==8.22.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jsbeautifier==1.15.1
json5==0.9.25
jupyter_client==8.6.0
jupyter_core==5.7.1
limits==3.13.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mdurl==0.1.2
nest-asyncio==1.6.0
numpy==1.26.4
openpyxl==3.1.2
ordered-set==4.1.0
packaging==24.0
pandas==2.2.1
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
Pygments==2.17.2
pymongo==4.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
regex==2023.12.25
requests==2.31.0
rich==13.7.1
six==1.16.0
stack-data==0.6.3
tomli==2.0.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
waitress==3.0.0
wcwidth==0.2.13
Werkzeug==3.0.1
wrapt==1.16.0
WTForms==2.3.3
